+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/Proxy/NativeProxyTest.php
@@ -17,6 +17,8 @@ use Symfony\Component\HttpFoundation\Session\Storage\Proxy\NativeProxy;
 /**
  * Test class for NativeProxy.
  *
+ * @group legacy
+ *
  * @author Drak <drak@zikula.org>
  */
 class NativeProxyTest extends TestCase